pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "research-os"
version = "0.1.0"
description = "Backend utilities for research manuscript workflows"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.116.0,<1.0.0",
    "httpx>=0.28.0,<1.0.0",
    "openai==2.21.0",
    "python-dotenv>=1.0.0,<2.0.0",
    "uvicorn[standard]>=0.30.0,<1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0,<9.0.0",
    "pytest-asyncio>=0.24.0,<1.0.0",
    "ruff>=0.8.0,<1.0.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F"]

[tool.ruff.format]
